Remove 9-byte header from output image
The original source has a 9-byte file header, which is present at the
start of the first data sector, but doesn't form part of the actual file
content. This was probably used to allow the image to be written to disk
starting cyl 4 head 0 sector 1 using raw sector writes.

Removing the header gives the true 10000 byte DOS image, which is bootable
when placed as the first file on a disk.